Time: 0 ms (100%), Space: 113.7 MB (61.93%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     double findMaxAverage(vector<int>& nums, int k) {
-        int current = 0, sum = 0;
+        double current = 0, sum = 0;
         for (int i = 0; i < k; ++i)
             current += nums[i];
         sum = current;
@@ -10,6 +10,6 @@ public:
             current += nums[i];
             sum = max(current, sum);
         }
-        return (double)sum / k;
+        return sum / k;
     }
 };
